Build extended

+++ build.sh
@@ -1,24 +1,46 @@
 #!/bin/sh
 # Build of the StaTIX
 # The only optional parameter is the jar output dir
-#
-# ./build.sh [-p] [<outdir>]
-# -p,--pack - build the tarball besides the executables
 
+USAGE="$0 [-p] [-c] [<outdir>]
+  -p,--pack - build the tarball besides the executables
+  -c,--classes  - retain classes after the build, useful for the frequent
+    modification and recompilation of some files.
+    
+    Compilation or the single file (.java to .class):
+    $ javac -cp lib/\*:src -d classes/ src/info/exascale/SimWeighted/main.java
+"
+# Extract the leading "-" if any:  ${1%%[^-]*}
 # Process input options
-TARBALL=0
-case $1 in
--p|--pack)
-	TARBALL=1
-	shift  # Shift the arguments
-	;;
--*)
-	echo "Invalid option specified, usage:"\
-		"\n  ./build.sh [-p] [<outdir>=\".\"]"\
-		"-p,--pack - build the tarball besides the executables"
-	exit 1
-	;;
-esac
+TARBALL=0  # Make tarball
+DELCLS=1  # Delete the classes after the jar building
+
+while [ -n "${1}" ]
+do
+	case $1 in
+	-p|--pack)
+		TARBALL=1
+		shift  # Shift the arguments
+		;;
+	-c|--classes)
+		DELCLS=0
+		shift
+		;;
+	-*)
+		printf "Error: Invalid option specified.\n\n$USAGE"
+		exit 1
+		;;
+	*)
+		# Check that only one output directory is specified
+		if [ -n "${2}"  ]
+		then
+			printf "Error: Too many parameters specified.\n\n$USAGE"
+			exit 1
+		fi
+		break
+		;;
+	esac
+done
 
 OUTDIR=${1:-.}  # Output directory for the executable package
 CLSDIR="$OUTDIR"/classes  # Classes output directory
@@ -45,9 +67,13 @@ then
 	echo Build failed, errcode: $?
 	exit $?
 fi
+
 # Remove the compiled classes
-echo Removing the \"$CLSDIR\"
-rm -rf "$CLSDIR"
+if [ $DELCLS -ne 0 ]
+then
+	echo Removing the \"$CLSDIR\"
+	rm -rf "$CLSDIR"
+fi
 
 # Copy requirements to the output dir
 if [ "$OUTDIR" != "." ]
